fix: move Cardradio to card folder and remove alpha property
