--- packages/@lightningjs/ui-components/src/components/CardRadio/index.d.ts
@@ -1,3 +1,0 @@
-import CardRadio, { CardRadioStyle } from './CardRadio';
-
-export { CardRadio as default, CardRadioStyle };
