--- packages/@lightningjs/ui-components/src/components/CardRadio/index.js
@@ -1,1 +1,0 @@
-export { default as default } from './CardRadio';
